third_party/blink/web_tests/platform/generic/fast/forms/email-idn-conversion-expected.txt

Tests ascii-unicode conversion of IDN email addresses on input type=email.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Values from UI
PASS i.value is "foo@xn--maana-pta.com"
PASS getUserAgentShadowTextContent(i) is "foo@mañana.com"
PASS i.validity.typeMismatch is false
PASS i.value is "mañana@foo.com"
PASS getUserAgentShadowTextContent(i) is "mañana@foo.com"
PASS i.validity.typeMismatch is true
PASS i.value is "mañana.com"
PASS getUserAgentShadowTextContent(i) is "mañana.com"
PASS i.validity.typeMismatch is true
Values from script
PASS i.value is "foo@xn--t8j.com"
PASS getUserAgentShadowTextContent(i) is "foo@お.com"
PASS i.validity.typeMismatch is false
PASS i.value is "xn--t8j@foo.com"
PASS getUserAgentShadowTextContent(i) is "xn--t8j@foo.com"
PASS i.validity.typeMismatch is false
PASS i.value is "xn--t8j.com"
PASS getUserAgentShadowTextContent(i) is "xn--t8j.com"
PASS i.validity.typeMismatch is true
PASS i.value is "foo@mañana.com"
PASS getUserAgentShadowTextContent(i) is "foo@mañana.com"
PASS i.validity.typeMismatch is true
Multiple addresses
PASS i.value is "foo@xn--maana-pta.com,bar@xn--t8j.com"
PASS getUserAgentShadowTextContent(i) is "foo@mañana.com, bar@お.com"
PASS i.validity.typeMismatch is false
PASS i.value is "bar@xn--t8j.com,foo@xn--maana-pta.com"
PASS getUserAgentShadowTextContent(i) is "bar@お.com,foo@mañana.com"
PASS i.validity.typeMismatch is false
PASS successfullyParsed is true

TEST COMPLETE
